Update ore_veins.json
The changes you've made to the mod requires that there's an array at the "ore" input field, but the default oreveins.json that's shipped with the mod isn't updated with those changes in mind, leading to a broken oreveins.json. 

Shipping it with this will spawn the example veins again.